requirements.txt
torch
torchvision
torchaudio
opencv-python
gradio
numpy
pillow
tqdm
huggingface_hub
python-dotenv
